pyproject.toml
[project]
name = "taekwonhandbook"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "gracjan", email = "gracjan.ziemianski@bcf-software.com" }
]
dependencies = [
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "django-cors-headers>=4.7.0",
    "drf-spectacular>=0.28.0",
    "django-filter>=25.1",
    "channels>=4.3.1",
    "channels-redis>=4.3.0",
    "celery>=5.5.3",
    "redis>=6.4.0",
    "psycopg[binary]>=3.2.9",
    "django-storages[s3]>=1.14.6",
    "meilisearch>=0.36.0",
    "uvicorn>=0.35.0",
    "gunicorn>=23.0.0",
    "pydantic>=2.11.7",
    "django-csp>=4.0",
    "python-json-logger>=3.3.0",
    "django-json-widget>=2.0.3",
    "rich>=14.1.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.5.0",
    "pytest>=8.2.0",
    "pytest-django>=4.8.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
pythonpath = ["backend"]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-select = ["I"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/taekwonhandbook"]
